Remove attributes data from save function for active filters
