+++ assets/js/blocks/active-filters/index.tsx
@@ -14,6 +14,7 @@ import edit from './edit';
 import metadata from './block.json';
 import { blockAttributes } from './attributes';
 import { Attributes } from './types';
+import deprecated from './deprecated';
 
 registerBlockType( metadata, {
 	icon: {
@@ -31,19 +32,13 @@ registerBlockType( metadata, {
 	edit,
 	// Save the props to post content.
 	save( { attributes }: { attributes: Attributes } ) {
-		const { className, displayStyle, heading, headingLevel } = attributes;
-		const data = {
-			'data-display-style': displayStyle,
-			'data-heading': heading,
-			'data-heading-level': headingLevel,
-		};
+		const { className } = attributes;
 
 		return (
 			<div
 				{ ...useBlockProps.save( {
 					className: classNames( 'is-loading', className ),
 				} ) }
-				{ ...data }
 			>
 				<span
 					aria-hidden
@@ -52,4 +47,5 @@ registerBlockType( metadata, {
 			</div>
 		);
 	},
+	deprecated,
 } );
